Got rid of path.py
